+++ mackup.py
@@ -481,7 +481,7 @@ SUPPORTED_APPS = {
 
 
 # Current version
-VERSION = '0.5.5'
+VERSION = '0.5.6'
 
 # Mode used to backup files to Dropbox
 BACKUP_MODE = 'backup'