+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Added support for LittleSnitch (thx @stechio)
 - Added support for OmniGraffle (thx @stecico)
 - Added support for SABnzbd (thx @stechico)
+- Added support for Skitch (thx @stechico)
 
 ## Mackup 0.5.5