+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Added support for OmniGraffle (thx @stecico)
 - Added support for SABnzbd (thx @stechico)
 - Added support for Skitch (thx @stechico)
+- Added support for FontExplorer X (thx @stechico)
 
 ## Mackup 0.5.5