+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Added support for FontExplorer X (thx @stechico)
 - Improved Transmission support (thx @stechico)
 - Added support for Lightroom 2, 3 and 4 (thx @stechico)
+- Fix Bundler’s synced dir (thx @atipugin)
 
 ## Mackup 0.5.5