+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## WIP
 
+## Mackup 0.5.6
+
 - Added support for LittleSnitch (thx @stechio)
 - Added support for OmniGraffle (thx @stecico)
 - Added support for SABnzbd (thx @stechico)